requirements.txt
mysql-connector-python
pg8000
sqlmodel
pydantic
sqlalchemy-utils
sqlalchemy-json
pytz
python-dotenv
pytest